rewrote Sha1Verify
